pyproject.toml
[tool.poetry]
name = "ingestion-benchmark"
authors = ["sflaviu <sflaviu96@gmail.com>"]
version = "1.0.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
tqdm = "^4.66.5"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
